requirements.txt
numpy
pandas
scipy
matplotlib
cartopy
shapely
xarray
netcdf4
pyyaml
requests
tqdm
